Document the public API (#101)
Add comments
Add missing package in Dockerfile
Add environment variable for wheel
Fix Python version in deb install
Version docs properly
Show all Python methods
Remove deprecated numpy.float
Fix bug with changing UID in container

+++ tests/workers/test_aks.py
@@ -33,7 +33,7 @@ class TestAks:
         for num in numbers:
             input_0 = amdinfer.InferenceRequestInput()
             input_0.name = "aks"
-            input_0.setFp32Data(np.asarray([num], np.float))
+            input_0.setFp32Data(np.asarray([num], np.float32))
             input_0.shape = [1]
             input_0.datatype = amdinfer.DataType.FP32
             request.addInputTensor(input_0)
